+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 ### Bug Fixes
 - The language switching is now more robust and preserves the original formatting from the style (`Map.setPrimaryLangage()`) (https://github.com/maptiler/maptiler-sdk-js/pull/134)
 
+### Others
+- Now able to GitHub action a beta on NPM from the GH release creation process
+- Updated GH action to v4
+
+
+
 ## 2.4.1
 ### Bug Fixes
 - The class `AJAXError` is now imported as part of the `maplibregl` namespace (CommonJS limitation from Maplibre GL JS) (https://github.com/maptiler/maptiler-sdk-js/pull/129)